Update the context parameters for the various VAT types.
The issue with the original parameters is that they could also trigger on the line-level. It can result in documents failing validation that should only get a warning (CR-561, due to line-level TaxTotal fields).

With this proposed commit, the rules will only fire on the document level, not the line level.